Binar X emulator: Hardwear mostly tested and started writing comments

+++ src/emulator.h
@@ -13,11 +13,24 @@
 
 namespace binarx_emulator {
 
-constexpr uint16_t kMaxPayloadDataLength = 500;
-constexpr uint16_t kDefaultCommunicationDelay = 200;
+constexpr uint16_t kMaxPayloadDataLength =
+    500; /**> Max payload data size allowed */
+constexpr uint32_t kDefaultCommunicationDelay =
+    200; /**> The time to wait for a communication transaction*/
 
+/**
+ * @brief Binar X Emulator to implement the Emulator functions
+ *
+ */
 class BinarXEmulator {
  public:
+  /**
+   * @brief Construct a new Binar X Emulator object with dependency injection
+   *
+   * @param spi_communication Pointer to the SPI implementation object
+   * @param uart_communication Pointer to the UART implementation object
+   * @param gpio_object Pointer to the GPIO implementation object
+   */
   BinarXEmulator(
       binarx_serial_interface::SerialCommunicationInterface* spi_communication,
       binarx_serial_interface::SerialCommunicationInterface* uart_communication,
@@ -25,9 +38,21 @@ class BinarXEmulator {
       : spi_comunication_(spi_communication),
         uart_comunication_(uart_communication),
         gpio_controller_(gpio_object){};
-
+  /**
+   * @brief Function for the SPI sequence
+   *
+   */
   void SpiRun();
+  /**
+   * @brief A function to toggle the Yellow LED to communicate that the Emulator
+   * is switched on
+   *
+   */
   void ToggleYellowLed();
+  /**
+   * @brief Function sequence for when the Button is pressed
+   *
+   */
   void ButtonPressed();
 
  private:
